requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
python-dotenv==1.0.0
notion-client==0.6.0
pydantic==1.10.10
httpx==0.24.1
